requirements.txt
fastapi
uvicorn[standard]
black
mypy

binance-connector
binance-futures-connector
pandas
matplotlib
plotly
numpy
requests
pydantic
aiohttp